revised per reviewer feedback
