--- examples/catalog/animated_list.iml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="FLUTTER_MODULE_TYPE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <excludeFolder url="file://$MODULE_DIR$/.idea" />
-      <excludeFolder url="file://$MODULE_DIR$/.pub" />
-      <excludeFolder url="file://$MODULE_DIR$/build" />
-      <excludeFolder url="file://$MODULE_DIR$/packages" />
-    </content>
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="Dart SDK" level="project" />
-    <orderEntry type="library" name="Dart SDK" level="application" />
-    <orderEntry type="library" name="Dart Packages" level="project" />
-  </component>
-</module>
